+++ index.html
@@ -12,6 +12,7 @@
       <header class="header">
         <a href="#registration" class="button-registration">Registration</a>
         <a href="#login" class="button-login">Login</a>
+        <a href="#logout" class="button-logout">Logout</a>
         <a href="#home">Home</a>
         <a href="#about">About</a>
         <a href="#contact">Contact</a>
